+++ Ejectify/View/StatusBar.swift
@@ -25,9 +25,4 @@ final class StatusBar {
             statusBarButton.image?.isTemplate = true
         }
     }
-
-    /// Rebuilds the status menu to reflect current state.
-    func refreshMenu() {
-        (statusItem.menu as? StatusBarMenu)?.refreshMenu()
-    }
 }